requirements.txt
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3        
idna==3.10
kiwisolver==1.4.8        
matplotlib==3.10.0       
numpy==2.2.1
packaging==24.2
panda==0.3.1
pillow==11.0.0
psutil==6.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-whois==0.9.5
requests==2.32.3
scapy==2.6.1
setuptools==75.6.0
six==1.17.0
urllib3==2.3.0